Fix MPRIS length field. Issue #1119

+++ mpd-interface/mpdstatus.h
@@ -72,8 +72,8 @@ struct MPDStatusValues {
     qint32 songId;
     qint32 nextSong;
     qint32 nextSongId;
-    qint32 timeElapsed;
-    qint32 timeTotal;
+    quint16 timeElapsed;
+    quint16 timeTotal;
     quint32 bitrate;
     quint32 samplerate;
     quint8 bits;
@@ -106,8 +106,8 @@ public:
     qint32 songId() const { return values.songId; }
     qint32 nextSong() const { return values.nextSong; }
     qint32 nextSongId() const { return values.nextSongId; }
-    qint32 timeElapsed() const { return values.timeElapsed; }
-    qint32 timeTotal() const { return values.timeTotal; }
+    quint16 timeElapsed() const { return values.timeElapsed; }
+    quint16 timeTotal() const { return values.timeTotal; }
     quint32 bitrate() const { return values.bitrate; }
     quint32 samplerate() const { return values.samplerate; }
     quint8 bits() const { return values.bits; }
@@ -118,8 +118,8 @@ public:
 
     // Cantata does not poll MPD for current position, but instead used a timer
     // This timer will update its value here - so this can be used elsewhere...
-    void setGuessedElapsed(qint32 v) { guessed=v; }
-    qint32 guessedElapsed() const { return guessed; }
+    void setGuessedElapsed(quint16 v) { guessed=v; }
+    quint16 guessedElapsed() const { return guessed; }
 
 public Q_SLOTS:
     void update(const MPDStatusValues &v);
@@ -132,7 +132,7 @@ private:
     MPDStatus& operator=(const MPDStatus& other);
 
 private:
-    qint32 guessed;
+    quint16 guessed;
     MPDStatusValues values;
 };
